LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/130-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected end of script

Harness Error (FAIL), message = SyntaxError: Unexpected end of script

FAIL scheduler: appending external script element to script  assert_array_equals: lengths differ, expected array ["inline script #1", "inline script #2", "end inline script #2", "end inline script #1", "external script #1"] length 5, got ["inline script #1", "end inline script #1", "external script #1"] length 3
